second chance for 0x02 project

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -3,7 +3,7 @@
 /*
  *
  *
- */void fibon(int n); /* Function protoypes for fibonacci */
+ */void fibon(unsigned int n); /* Function protoypes for fibonacci */
 
 /*
  *
@@ -22,10 +22,10 @@
  *
  * Function Definition here
  *
- */void fibon(int n)
+ */void fibon(unsigned int n)
 {
-	unsigned long long term1 = 1, term2 = 2;
-	unsigned long long i, next;
+	unsigned int term1 = 1, term2 = 2;
+	unsigned int i, next;
 
 	printf("1, 2, ");
